+++ types/byte_array.go
@@ -1,0 +1,87 @@
+package types
+
+import (
+	"encoding/base64"
+	"math/big"
+)
+
+// order=LittleEndian or BigEndian; length is byte num
+func StrIntToBinary(num, order string, length int, signed bool) string {
+	bigNum := new(big.Int)
+	bigNum.SetString(num, 10)
+	if !signed {
+		res := bigNum.Bytes()
+		if len(res) < length {
+			res = append(make([]byte, length-len(res)), res...)
+		}
+		if order == "LittleEndian" {
+			for i, j := 0, len(res)-1; i < j; i, j = i+1, j-1 {
+				res[i], res[j] = res[j], res[i]
+			}
+		}
+		if length > 0 {
+			res = res[len(res)-length:]
+		}
+		return string(res)
+	}
+
+	flag := bigNum.Cmp(big.NewInt(0))
+	if flag == 0 {
+		if length <= 0 {
+			length = 1
+		}
+		return string(make([]byte, length))
+	}
+
+	bigNum = bigNum.SetBytes(bigNum.Bytes())
+	bs := bigNum.Bytes()
+
+	if len(bs) < length {
+		bs = append(make([]byte, length-len(bs)), bs...)
+	}
+
+	upperBs := make([]byte, len(bs))
+	upperBs[0] = byte(0x80)
+	upper := new(big.Int)
+	upper.SetBytes(upperBs)
+	if flag > 0 {
+		upper = upper.Sub(upper, big.NewInt(1))
+	}
+
+	if bigNum.Cmp(upper) > 0 {
+		bs = append(make([]byte, 1), bs...)
+	}
+
+	if flag < 0 {
+		modBs := make([]byte, len(bs)+1)
+		modBs[0] = byte(0x01)
+		mod := new(big.Int)
+		mod.SetBytes(modBs)
+		bs = mod.Sub(mod, bigNum).Bytes()
+	}
+	if length > 0 {
+		bs = bs[len(bs)-length:]
+	}
+	if order == "LittleEndian" {
+		for i, j := 0, len(bs)-1; i < j; i, j = i+1, j-1 {
+			bs[i], bs[j] = bs[j], bs[i]
+		}
+	}
+	return string(bs)
+}
+
+// convertBinaryValue handles base64 encoding for binary data without logical/converted types.
+func convertBinaryValue(val any) any {
+	if val == nil {
+		return nil
+	}
+
+	switch v := val.(type) {
+	case []byte:
+		return base64.StdEncoding.EncodeToString(v)
+	case string:
+		return base64.StdEncoding.EncodeToString([]byte(v))
+	}
+
+	return val
+}
