+++ types/uuid_test.go
@@ -1,0 +1,88 @@
+package types
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestConvertUUIDValue(t *testing.T) {
+	tests := []struct {
+		name     string
+		val      any
+		expected any
+	}{
+		{
+			name:     "nil_value",
+			val:      nil,
+			expected: nil,
+		},
+		{
+			name: "valid_uuid_bytes",
+			val: []byte{
+				0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
+				0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 0xef,
+			},
+			expected: "01234567-89ab-cdef-0123-456789abcdef",
+		},
+		{
+			name: "valid_uuid_string",
+			val: string([]byte{
+				0x6b, 0xa7, 0xb8, 0x10, 0x9d, 0xad, 0x11, 0xd1,
+				0x80, 0xb4, 0x00, 0xc0, 0x4f, 0xd4, 0x30, 0xc8,
+			}),
+			expected: "6ba7b810-9dad-11d1-80b4-00c04fd430c8",
+		},
+		{
+			name: "zero_uuid",
+			val: []byte{
+				0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+				0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			},
+			expected: "00000000-0000-0000-0000-000000000000",
+		},
+		{
+			name: "max_uuid",
+			val: []byte{
+				0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+				0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+			},
+			expected: "ffffffff-ffff-ffff-ffff-ffffffffffff",
+		},
+		{
+			name:     "invalid_length_short",
+			val:      []byte{0x01, 0x02, 0x03},
+			expected: []byte{0x01, 0x02, 0x03},
+		},
+		{
+			name: "invalid_length_long",
+			val: []byte{
+				0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08,
+				0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10,
+				0x11,
+			},
+			expected: []byte{
+				0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08,
+				0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10,
+				0x11,
+			},
+		},
+		{
+			name:     "non_binary_value",
+			val:      42,
+			expected: 42,
+		},
+		{
+			name:     "empty_bytes",
+			val:      []byte{},
+			expected: []byte{},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := ConvertUUIDValue(tt.val)
+			require.Equal(t, tt.expected, result)
+		})
+	}
+}
